requirements.txt
lxml==4.2.1
requests==2.18.4
urllib3==1.22
PyTelegramBotAPI==2.2.3
flask
psycopg2-binary
apscheduler

